+++ pkg/config/config.go
@@ -4,11 +4,12 @@ package config
 
 import (
 	"errors"
+	"io"
+	"reflect"
+
 	"github.com/davecgh/go-spew/spew"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
-	"io"
-	"reflect"
 )
 
 // RepoConfig stores config of each repo in a map
